// src/types/Listing.ts
export type Listing = {
    listing_id: string;
    provider_id: string;
    title: string;
    address: string;
    longitude: number;
    latitude: number;
    type: string;
    rate: number;
    description: string;
    status: string;
    image: string;
    created_at: string;
    updated_at: string;
};